requirements.txt
pytorch==1.6
fairseq==0.9
transformers==2.9